Add new link type for zip filings to filinghistory
New functionality has been added to the filing history
template to now allow the downloading of zip file types.

Currently we identify zip accounts via their unique page
count of 9999. This was decided as an initial implentation
which will be properly implenented with a more elegant
soltion in the near future.