Refactor code to reduce amount of document API calls
We've added more criteria for each filing to meet before we
call the document API. This will lighten the load on API calls.

logging has also been added to various points in the flow so that
it's easier for debugging in the future.

Null check has been added to the resources array returned to avoid
calling a null object.