Fix temporary implementation of get_content_tye sub
Before this fix, the sub routine would only return the first
content_type that it found in the resrouces array. This was
only temporary to meet a strict deadline.

This change means that the sub routine will now loop through
the resources array that is returned from doc api
and attempt to find the content_type we want and set it
correctly.