Fix GOV.UK Tabs breaking when running JavaScript
With the new rebrand work we started to run the GOV.UK Frontend
JavaScript code and noticed that as is, when JavaScript initialises
it breaks our usage of tabs.
We have not had a problem until now because we've not initalised JavaScript,
which could be considered a bug.

The GOV.UK Tabs component is typically used to change
content in a single page without refreshing the page.
We are using it to navigate between pages with full page refreshes.
This means we do not need to initialise the JavaScript for the
GOV.UK Tabs component on this page.

The `data-module` attribute tells GOV.UK Frontend to initialise the JavaScript for a component,
removing this attribute makes the component HTML only.